Deprecate redundant interfaces
Many of the interfaces are exact duplicates of ones from GLSurfaceView
and, as far as I can tell, add no value. While integrating the latest
code into the downstream Rajawali framework, I found these interfaces to
be a nuisance. Our classes, which already implement the GLSurfaceView
interfaces, were not considered valid to the GLWallpaperService.

GLWallpaperService has been modified to accept the GLSurfaceView
interfaces. To maintain backwards compatibility, the redundant
GLWallpaperService interfaces have been changed to empty extensions of
their GLSurfaceView counterparts, but marked as deprecated. The downside
of keeping the interfaces for backwards compatibility is that
"GLSurfaceView." had to be prefixed onto many declarations.