Replace redundant interfaces with GLSurfaceView
Many of the interfaces are exact duplicates of ones from GLSurfaceView
and, as far as I can tell, add no value. While integrating the latest
code into the downstream Rajawali framework, I found these interfaces to
be a nuisance. Our classes, which already implement the GLSurfaceView
interfaces, were not considered valid to the GLWallpaperService.

GLWallpaperService has been modified to accept the GLSurfaceView
interfaces. The local duplicates have been removed. This will break
backward compatibility for projects that use the local variant, but the
fix is easy: replace the "GLWallpaperService." prefix with a
"GLSurfaceView." prefix or tweak the import statements.